+++ simulation/traffic_simulator/test/src/test_math.cpp
@@ -56,7 +56,7 @@ TEST(Math, HermiteCurve1)
   EXPECT_DOUBLE_EQ(curve.getLength(), 1);
   EXPECT_DOUBLE_EQ(curve.getPoint(0.5, false).x, 0.5);
   EXPECT_DOUBLE_EQ(curve.getTangentVector(0.5, false).x, 1);
-  EXPECT_DOUBLE_EQ(curve.getMaximu2DCurvature(), 0);
+  EXPECT_DOUBLE_EQ(curve.getMaximum2DCurvature(), 0);
   geometry_msgs::msg::Point p;
   p.x = 0.1;
   p.y = 0;
@@ -87,7 +87,7 @@ TEST(Math, HermiteCurve2)
   EXPECT_DOUBLE_EQ(curve.getLength(), 1);
   EXPECT_DOUBLE_EQ(curve.getPoint(0.5, false).x, 0.5);
   EXPECT_DOUBLE_EQ(curve.getTangentVector(0.5, false).x, 1);
-  EXPECT_DOUBLE_EQ(curve.getMaximu2DCurvature(), 0);
+  EXPECT_DOUBLE_EQ(curve.getMaximum2DCurvature(), 0);
   geometry_msgs::msg::Point p;
   p.x = 0.1;
   p.y = 0;